Tweaking off-stream to add collision detection.
Uses 3 lasers forward too now, to find the normal of what's ahead- if it's
too horizontal it assumes it's an obstacle and it overrides the
goal-seeking steering, bending it left or right depending on which way that
normal is facing.  Works to avoid convex objects but has some
failure modes with concave objects like the KSC buildings that have inner
corners.  Instead of seeking to go around them it ends up seeking into the
corner and getting stuck.  This is okay becuase the sorts of rocks it
was meant to avoid don't have concave corners like the test KSC buildings do.