Found a bug in ullage logic for land_it_vary.
The bug happened when starting the suicide burn by pushing RCS
forward for ullage actually has a big enough effect to
cause the predicted stopping point to move up well above
the ground before the ullage step is over.  When this
happened, the script fell into an "else" where it set the
throttle to 0.001, thinking it was throtting down after the
burn started, when actually it was throttling up to 0.001
prior to starting the burn.  This caused an attempt to
ignite the engine before the ullage step was done.