Node burn now seeks AP or PE to extend burn.
If the burn is from a maneuver node with prograde
or retrograde content, it will now pick either the Ap
or the Pe to monitor and decide the burn is done only
when BOTH the dv_to_go is <= 0 AND the target Ap or Pe
has been reached.  If the dV_to_go is done, but the
goal isn't reached, it still burns a bit longer till it is.

It picks the Ap or Pe as the goal depending on where the burn
takes place. (it picks whichever is on the opposite side of
the orbit from where the burn happens.)