Handle edge case of incl burn in node script.
There is an edge case where an inclination change burn
that is along a vector that is part retrograde and
part (anti)normal will reach low Pe but just a hair too
shy of low enough to to end the script, so it starts
rising the Pe after it should have stopped the burn.
This adds the check to see if the "direction" of the
Pe or Ap changes has flipped and if it has, stop the
burn.