rover diverting sideways now also considers accel.
The rover diverts to a sideways movement for a while
if it thinks forward momentum is stalled.  But previously
it assumed it was stalled if it was going slower than
desired, even if it's actually still accellerating up to
speed.  Now it only consideres itself stalled if the
speed is both slow *and* decellerating.  If the speed is
increasing, then it stays with the forward motion assuming
it will get better.