fix: wantclimb should be to aimpoint, not to waypoint.
It was setting the wantClimb to climb to the waypoint at the
end of the (red) course line, instead of climbing to the aimpoint
(yellow) line, thus it was perpetually below the red course line
and staying that way on landings.

(Also I see some confusing clash in termiology in the script's
variable names as to what an "aim point" is - some variables
use that for the waypoint targets, others for the yellow aimpoint
target.  Maybe need a future refactor to rename them to make it
clearer.)