pilot tweak wheelsteering & stager bugfix double stage.
Pilot: Wheelsteer needed to be milder.
Stager: There was a "race condition" where the game no longer
instantly decouples everything when you say "stage".  If it
decides not to finish the act of staging, you still have some engines
attached for a physics tick that should have staged off.  Added
some checks so the detatched engines don't cause the stager to falsely
think there's a need for an additional STAGE command.