fix ullage check getting stuck forever.
Ullage check was getting stuck forever when the staging event
just after an engine died was NOT the staging event that
decoupled that engine.

Example:

If the staging from first stage to second stage goes:
First - decouple fairing.
Second - engage RCS blocks.
Third - detatch lower stage.
Fourth - Activate next engine.

Then the problem is that the flamed out engine from the prev
stage is stil attached the first time it ran the ullage check,
and the existence of that still attached engine made the ullage
check return false every time - because that engine had bad
pressure on account of its fuel tank being emptied, not on
account of how many Gee's it was pulling.  The fix is to realize
that flamed-out engines need to be bypassed in the ullage check,
as they will always fail ullage by having empty fuel and don't count
as the next set of engines you are about to ignite.