Fix bug where launch aborted when aiming down pushing Pe.
If in a launch and the Pe is already high enough that we won't fall in,
it was still aborting the launch anyway as if we were falling in,
thinking we were "falling" because of the negative vertical speed when
the Pe was being pushed out ahead of us by burning in an nose-down
orientation for circularization.

Fix: if Pe is high so we won't be falling in, don't abort even if
"going downward".